pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nicescholia"
version = "0.0.1"
description = "nicegui based Scholia"
authors = [
  {name = "Wolfgang Fahl", email = "wf@bitplan.com"}
]
readme = "README.md"
license = "Apache-2.0"
dependencies = [
  # https://github.com/WolfgangFahl/pybasemkit
  "pybasemkit>=0.1.6",
  # https://github.com/WolfgangFahl/nicegui_widgets
  "ngwidgets>=0.30.3",
  # https://github.com/WolfgangFahl/snapquery
  "snapquery>=0.2.4",
  # https://github.com/encode/httpx
  "httpx>=0.27.0",
]
requires-python = ">=3.11"

[project.urls]
Home = "https://github.com/WolfgangFahl/nicescholia"

[project.scripts]
nicescholia = "nscholia.cmd:main"

[tool.hatch.build.targets.wheel]
only-include = ["nscholia", "nscholia_examples"]
